Formatted statistics.js to my liking.
Mostly spaces around `=` and after `;` in for loops.

Line 31 seems weird to me, I don't know what that's supposed to be doing.
Commented on that in line 30.

New variables should always be declared with var so that their scope is
set to be local to the function that they belong to. If they're not
initialized with var, then they're considered global variables, and that's
messy.