Attempting to pass around WIDTH variable.
Instead of relying on the WIDTH variable being global, I'm trying to pass
it to the functions that use it... because I don't like global variables.
After writing that out, it isn't a particularly compelling argument to
make such a change, but hopefully the changes to come will make this
dislike of globals reasonable.

This changes breaks the page. I'm not sure what to do about the jQuery
based UI functions that make use of WIDTH, so I'll leave those UI
functions alone for now, and come back to them after the Graph and
Population functions are reigned in.